+++ backend/src/main/java/com/hethongtruongthpt/controller/LopHocController.java
@@ -23,7 +23,7 @@ public class LopHocController {
 	}
 
 	@GetMapping("/{id}")
-	public ResponseEntity<ApiResponse<LopHoc>> getById(@PathVariable Long id) {
+	public ResponseEntity<ApiResponse<LopHoc>> getById(@PathVariable Integer id) {
 		return ResponseEntity.ok(ApiResponse.ok(lopHocService.getById(id)));
 	}
 
@@ -33,12 +33,12 @@ public class LopHocController {
 	}
 
 	@PutMapping("/{id}")
-	public ResponseEntity<ApiResponse<LopHoc>> update(@PathVariable Long id, @RequestBody LopHoc lopHoc) {
+	public ResponseEntity<ApiResponse<LopHoc>> update(@PathVariable Integer id, @RequestBody LopHoc lopHoc) {
 		return ResponseEntity.ok(ApiResponse.ok(lopHocService.update(id, lopHoc)));
 	}
 
 	@DeleteMapping("/{id}")
-	public ResponseEntity<ApiResponse<Object>> delete(@PathVariable Long id) {
+	public ResponseEntity<ApiResponse<Object>> delete(@PathVariable Integer id) {
 		lopHocService.delete(id);
 		return ResponseEntity.ok(ApiResponse.ok("Xóa thành công", null));
 	}
