+++ backend/src/main/java/com/hethongtruongthpt/entity/DangTaoId.java
@@ -8,24 +8,24 @@ import java.io.Serializable;
 @Embeddable
 public class DangTaoId implements Serializable {
     @Column(name = "ID_USER")
-    private Long userId;
+    private Integer userId;
 
     @Column(name = "ID_THONGBAO")
-    private Long thongBaoId;
+    private Integer thongBaoId;
 
-    public Long getUserId() {
+    public Integer getUserId() {
         return userId;
     }
 
-    public void setUserId(Long userId) {
+    public void setUserId(Integer userId) {
         this.userId = userId;
     }
 
-    public Long getThongBaoId() {
+    public Integer getThongBaoId() {
         return thongBaoId;
     }
 
-    public void setThongBaoId(Long thongBaoId) {
+    public void setThongBaoId(Integer thongBaoId) {
         this.thongBaoId = thongBaoId;
     }
 }
